requirements/requirements.txt
appdirs==1.4.4
astroid==2.3.3
colorama==0.4.3
distlib==0.3.1
Django==2.2
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.2
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.4.4
pytz==2020.4
six==1.14.0
sqlparse==0.4.1
virtualenv==20.2.1
wrapt==1.11.2
